__tests__/internals/index.ts -> __tests__/internals/sentry.ts
